Factor out main() into separate file. Refactor some python to make it compatible with transcrypt
